tools: Split openpgp-tool's prettify_* functions into another file.
No change to source code other than the relocation.
In preparation for unit testing.
